refactor: navbarProfile mypage 폴더로 이동
