+++ src/pages/MyPage/NavbarProfile/index.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Nav, IconContainer } from './styles';
 import { Link } from 'react-router-dom';
 import settingIcon from '../../assets/default/setting.svg';
-import { StyledText } from '../../components/Text/StyledText';
-import theme from '../../styles/theme';
+import { StyledText } from '../../../components/Text/StyledText';
+import theme from '../../../styles/theme';
 
 const NavbarProfile: React.FC = () => {
 	return (
